Add files for bouncing text animation
This code creates an animation which breaks a text string into characters and
words, "bouncing" them off the edges of the browser window.  We start with
the first character.  When it hits the right wall, the second character emerges
from it, and so forth.  Eventually a word emerges.  Each new text item is given
a random color and a random speed (the latter so texts don't clump).

The file collision_test.js is a sketch for a further enhancement.  It would be
nice to have the letters and words bounce off each other rather than simply the
walls.  For me, this enhancement is a test of how fast Javascript is, whether
this sort of thing ought to be done in a compiled language.  With 27 elements
currently in play, that's a lot of objects that have to be constantly monitored
for collisions.

A tricky issue will be not to register a collision when the letters are
initially on top of each other at the right wall.